# speaksafe-backend/docs/openapi.yaml
openapi: 3.1.1
info:
  title: SpeakSafe API
  version: 1.0.0
  description: |
    Anonymous reporting system for boarding schools with multi-school support.

    ## Key Features
    - **School Registration** - Schools submit registration requests (pending approval)
    - **Multi-School Support** - Each school has isolated data
    - **Three User Roles** - System Admin, School Admin, School Staff
    - **Anonymous Reporting** - Submit reports without revealing identity
    - **Status Tracking** - Check report status using reference codes
    - **Authority Dashboard** - Manage and investigate reports

    ## User Roles
    | Role | Who | Access |
    |------|-----|--------|
    | **System Admin** | SpeakSafe Team | Full platform access |
    | **School Admin** | School Authority | Full school access |
    | **School Staff** | Staff under School Admin | Limited access (assigned reports only) |

    ## Authentication
    - JWT tokens stored in HTTP-Only Cookies
    - Access Token expires in 15 minutes
    - Refresh Token expires in 7 days
    - Use cookie `accessToken` or `Authorization: Bearer <token>` header

  contact:
    name: SpeakSafe Backend Team
    email: circoquantumcircle@gmail.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://speaksafe-puzj.onrender.com/api/v1
    description: Production Server
  - url: http://localhost:5000/api/v1
    description: Development Server
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
      description: JWT access token stored in HTTP-only cookie
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token in Authorization header
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
    Tokens:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token (expires in 15 minutes)
        refreshToken:
          type: string
          description: JWT refresh token (expires in 7 days)
...
    SubmitRegistrationRequest:
      type: object
      required:
        - schoolName
        - domain
        - adminName
        - adminEmail
        - adminPassword
      properties:
        schoolName:
          type: string
          minLength: 2
          maxLength: 100
          example: "St. Mary's School"
        domain:
          type: string
          minLength: 2
          maxLength: 100
          example: "stmarys.edu"
          description: "Email domain for school"
        adminName:
          type: string
          minLength: 2
          maxLength: 100
          example: "John Smith"
        adminEmail:
          type: string
          format: email
          example: "admin@stmarys.edu"
        adminPassword:
          type: string
          minLength: 8
          example: "AdminPassword123!"
          description: "Must contain uppercase, lowercase, and number"
        address:
          type: string
          maxLength: 200
          optional: true
        phone:
          type: string
          maxLength: 20
          optional: true
        email:
          type: string
          format: email
          optional: true
        website:
          type: string
          format: uri
          optional: true
    RegistrationResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            id:
              type: string
            schoolName:
              type: string
            domain:
              type: string
            adminEmail:
              type: string
            status:
              type: string
              enum: [pending, approved, rejected]
            submittedAt:
              type: string
              format: date-time
    ReviewRegistrationRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [approved, rejected]
        reviewNotes:
          type: string
          maxLength: 500
    RegistrationStatsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            total:
              type: number
            pending:
              type: number
            approved:
              type: number
            rejected:
              type: number

    # ====== School Schemas ======
...
    Pagination:
      type: object
      properties:
        total:
          type: number
        page:
          type: number
        limit:
          type: number
        totalPages:
          type: number
...
paths:
...
  /registrations:
    post:
      tags:
        - Registrations
      summary: Submit School Registration
      description: Submit a school registration request (Public)
      operationId: submitRegistration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitRegistrationRequest"
      responses:
        "201":
          description: Registration submitted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationResponse"
        "409":
          description: Domain or email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      tags:
        - Registrations
      summary: Get Registrations
      description: Get all school registrations (System Admin only)
      operationId: getRegistrations
      security:
        - cookieAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: number
            default: 1
        - name: limit
          in: query
          schema:
            type: number
            default: 20
      responses:
        "200":
          description: Registrations retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      registrations:
                        type: array
                        items:
                          $ref: "#/components/schemas/RegistrationResponse"
                      pagination:
                        $ref: "#/components/schemas/Pagination"
                      stats:
                        $ref: "#/components/schemas/RegistrationStatsResponse"
        "401":
          description: Not authenticated
        "403":
          description: Insufficient permissions
  /registrations/stats:
    get:
      tags:
        - Registrations
      summary: Get Registration Stats
      description: Get statistics for school registrations (System Admin only)
      operationId: getRegistrationStats
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Statistics retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationStatsResponse"
        "401":
          description: Not authenticated
        "403":
          description: Insufficient permissions
  /registrations/{id}/review:
    put:
      tags:
        - Registrations
      summary: Review Registration
      description: Approve or reject a school registration (System Admin only)
      operationId: reviewRegistration
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Registration ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewRegistrationRequest"
      responses:
        "200":
          description: Registration reviewed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      status:
                        type: string
                        enum: [approved, rejected]
                      school:
                        type: object
                        properties:
                          id:
                            type: string
                          name:
                            type: string
                          domain:
                            type: string
                      admin:
                        type: object
                        properties:
                          id:
                            type: string
                          email:
                            type: string
                          name:
                            type: string
                      tokens:
                        $ref: "#/components/schemas/Tokens"
        "404":
          description: Registration not found
        "400":
          description: Registration already reviewed

  # ====== SCHOOL ROUTES ======
...